HTML {htmltools}	R Documentation

Mark Characters as HTML

Description

Marks the given text as HTML, which means the tag functions will know not to perform HTML escaping on it.

Usage

HTML(text, ..., .noWS = NULL)

Arguments

text
The text value to mark with HTML

...
Any additional values to be converted to character and concatenated together

.noWS
Character vector used to omit some of the whitespace that would normally be written around this HTML. Valid options include before, after, and outside (equivalent to before and end).

Value

The input text, but marked as HTML.

Examples

el <- div(HTML("I like <u>turtles</u>"))
cat(as.character(el))

[Package htmltools version 0.5.2 Index]