as.tags {htmltools}	R Documentation

Convert a value to tags

Description

An S3 method for converting arbitrary values to a value that can be used as the child of a tag or tagList. The default implementation simply calls as.character().

Usage

as.tags(x, ...)

Arguments

x
Object to be converted.

...
Any additional parameters.

[Package htmltools version 0.5.6.1 Index]