makeDependencyRelative {htmltools}	R Documentation

Make an absolute dependency relative

Description

Change a dependency's absolute path to be relative to one of its parent directories.

Usage

makeDependencyRelative(dependency, basepath, mustWork = TRUE)

Arguments

dependency
A single HTML dependency with an absolute path.

basepath
The path to the directory that dependency should be made relative to.

mustWork
If TRUE and dependency does not point to a directory on disk (but rather a URL location), an error is raised. If FALSE then non-disk dependencies are returned without modification.

Value

The dependency with its src value updated to the new location's relative path.

If baspath did not appear to be a parent directory of the dependency's directory, an error is raised (regardless of the value of mustWork).

See Also

copyDependencyToDir()

[Package htmltools version 0.5.6.1 Index]