resolveDependencies {htmltools}	R Documentation

Resolve a list of dependencies

Description

Given a list of dependencies, removes any redundant dependencies (based on name equality). If multiple versions of a dependency are found, the copy with the latest version number is used.

Usage

resolveDependencies(dependencies, resolvePackageDir = TRUE)

Arguments

dependencies
A list of htmlDependency() objects.

resolvePackageDir
Whether to resolve the relative path to an absolute path via system.file() when the package attribute is present in a dependency object.

Value

dependencies A list of htmlDependency() objects with redundancies removed.

[Package htmltools version 0.5.6.1 Index]