HTML Tag Object

tag() creates an HTML tag definition. Note that all of the valid HTML5 tags are already defined in the tags() environment so these functions should only be used to generate additional tags. tagAppendChild() and tagList() are for supporting package authors who wish to create their own sets of tags; see the contents of bootstrap.R for examples.

tagList(...)

tagAppendAttributes(tag, ...)

tagHasAttribute(tag, attr)

tagGetAttribute(tag, attr)

tagAppendChild(tag, child)

tagAppendChildren(tag, ..., list = NULL)

tagSetChildren(tag, ..., list = NULL)

tag(`_tag_name`, varArgs, .noWS = NULL, .renderHook = NULL)

Arguments

...
Unnamed items that comprise this list of tags.

tag
A tag to append child elements to.

attr
The name of an attribute.

child
A child element to append to a parent tag.

list
An optional list of elements. Can be used with or instead of the ... items.

_tag_name
HTML tag name

varArgs
List of attributes and children of the element. Named list items become attributes, and unnamed list items become children. Valid children are tags, single-character character vectors (which become text nodes), and raw HTML (see HTML()). You can also pass lists that contain tags, text nodes, and HTML.

.noWS
Character vector used to omit some of the whitespace that would normally be written around this tag. Valid options include before, after, outside, after-begin, and before-end. Any number of these options can be specified.

.renderHook
A function (or list of functions) to call when the tag is rendered. This function should have at least one argument (the tag) and return anything that can be converted into tags via as.tags(). Additional hooks may also be added to a particular tag via tagAddRenderHook().

Value

An HTML tag object that can be rendered as HTML using as.character().

Examples

tagList(tags$h1("Title"),
        tags$h2("Header text"),
        tags$p("Text here"))

#> <h1>Title</h1>
#> <h2>Header text</h2>
#> <p>Text here</p>

# Can also convert a regular list to a tagList (internal data structure isn't
# exactly the same, but when rendered to HTML, the output is the same).
x <- list(tags$h1("Title"),
          tags$h2("Header text"),
          tags$p("Text here"))
tagList(x)

#> <h1>Title</h1>
#> <h2>Header text</h2>
#> <p>Text here</p>

# suppress the whitespace between tags
oneline <- tag("span",
  tag("strong", "Super strong", .noWS="outside")
)
cat(as.character(oneline))

#> <span name="strong" children="Super strong" .noWS="outside"></span>

# At print time, turn an h1 into an h2 tag
h <- tags$h1("Example", .renderHook = function(x) {
  x$name <- "h2"
  x
})
h

#> <h2>Example</h2>

oneline <- tag("span",
  tag("strong", "Super strong", .noWS="outside")
)
cat(as.character(oneline))

#> <span name="strong" children="Super strong" .noWS="outside"></span>