tagAppendAttributes {htmltools}	R Documentation

Append tag attributes

Description

Append (tagAppendAttributes()), check existence (tagHasAttribute()), and obtain the value (tagGetAttribute()) of HTML attribute(s).

Usage

tagAppendAttributes(tag, ..., .cssSelector = NULL)

tagHasAttribute(tag, attr)

tagGetAttribute(tag, attr)

Arguments

tag
a tag object.

...
a collection of attributes.

.cssSelector
A character string containing a CSS selector for targeting particular (inner) tags of interest. At the moment, only a combination of type (e.g, div), class (e.g., .my-class), id (e.g., ⁠#myID⁠), and universal (*) selectors within a given simple selector is supported. Note, if .cssSelector is used, the returned tags will have their ⁠$children⁠ fields flattened to a single list() via tagQuery().

attr
The name of an attribute.

See Also

tagAppendChildren(), tagQuery()

Examples

html <- div(a())
tagAppendAttributes(html, class = "foo")
tagAppendAttributes(html, .cssSelector = "a", class = "bar")

tagHasAttribute(div(foo = "bar"), "foo")
tagGetAttribute(div(foo = "bar"), "foo")

[Package htmltools version 0.5.6.1 Index]