browsable {htmltools}	R Documentation

Make an HTML object browsable

Description

By default, HTML objects display their HTML markup at the console when printed. browsable can be used to make specific objects render as HTML by default when printed at the console.

Usage

browsable(x, value = TRUE)

is.browsable(x)

Arguments

x
The object to make browsable or not.

value
Whether the object should be considered browsable.

Details

You can override the default browsability of an HTML object by explicitly passing browse = TRUE (or FALSE) to the print function.

Value

browsable returns x with an extra attribute to indicate that the value is browsable.

is.browsable returns TRUE if the value is browsable, or FALSE if not.

[Package htmltools version 0.5.2 Index]