findDependencies {htmltools}	R Documentation

Collect attached dependencies from HTML tag object

Description

Walks a hierarchy of tags looking for attached dependencies.

Usage

findDependencies(tags, tagify = TRUE)

Arguments

tags
A tag-like object to search for dependencies.

tagify
Whether to tagify the input before searching for dependencies.

Value

A list of htmlDependency() objects.

[Package htmltools version 0.5.2 Index]