htmlEscape {htmltools}	R Documentation

Escape HTML entities

Description

Escape HTML entities contained in a character vector so that it can be safely included as text or an attribute value within an HTML document

Usage

htmlEscape(text, attribute = FALSE)

Arguments

text
Text to escape

attribute
Escape for use as an attribute value

Value

Character vector with escaped text.

[Package htmltools version 0.5.2 Index]