html_print {htmltools}	R Documentation

Implementation of the print method for HTML

Description

Convenience method that provides an implementation of the base::print() method for HTML content.

Usage

html_print(
  html,
  background = "white",
  viewer = getOption("viewer", utils::browseURL)
)

Arguments

html
HTML content to print

background
Background color for web page

viewer
A function to be called with the URL or path to the generated HTML page. Can be NULL, in which case no viewer will be invoked.

Value

Invisibly returns the URL or path of the generated HTML page.

[Package htmltools version 0.5.2 Index]