renderDependencies {htmltools}	R Documentation

Create HTML for dependencies

Description

Create the appropriate HTML markup for including dependencies in an HTML document.

Usage

renderDependencies(
  dependencies,
  srcType = c("href", "file"),
  encodeFunc = urlEncodePath,
  hrefFilter = identity
)

Arguments

dependencies
A list of htmlDependency objects.

srcType
The type of src paths to use; valid values are file or href.

encodeFunc
The function to use to encode the path part of a URL. The default should generally be used.

hrefFilter
A function used to transform the final, encoded URLs of script and stylesheet files. The default should generally be used.

Value

An HTML() object suitable for inclusion in the head of an HTML document.

[Package htmltools version 0.5.2 Index]