resolveDependencies {htmltools}	R Documentation

Resolve a list of dependencies

Description

Given a list of dependencies, removes any redundant dependencies (based on name equality). If multiple versions of a dependency are found, the copy with the latest version number is used.

Usage

resolveDependencies(dependencies)

Arguments

dependencies
A list of htmlDependency objects.

Value

dependencies A list of htmlDependency objects with redundancies removed.

[Package htmltools version 0.3.5 Index]