subtractDependencies {htmltools}	R Documentation

Subtract dependencies

Description

Remove a set of dependencies from another list of dependencies. The set of dependencies to remove can be expressed as either a character vector or a list; if the latter, a warning can be emitted if the version of the dependency being removed is later than the version of the dependency object that is causing the removal.

Usage

subtractDependencies(dependencies, remove, warnOnConflict = TRUE)

Arguments

dependencies
A list of htmlDependency() objects from which dependencies should be removed.

remove
A list of htmlDependency() objects indicating which dependencies should be removed, or a character vector indicating dependency names.

warnOnConflict
If TRUE, a warning is emitted for each dependency that is removed if the corresponding dependency in remove has a lower version number. Has no effect if remove is provided as a character vector.

Value

A list of htmlDependency() objects that don't intersect with remove.

[Package htmltools version 0.5.2 Index]