Perform jQuery-like alterations on tags

tagQuery(tags)

Arguments

tags
Any standard tag object or tagList(). If a list() or tagList() is provided, a tagList() will be returned when calling $asTags().

Value

A tagQuery() object. The tag supplied will be considered the root object. At the time of initialization, the root is also considered the single selected item. If any selections are made, the selected elements will be updated.

Details

This function is VERY experimental. The result api will most likely change. Use at your own risk.

tagQuery() is modeled after jQuery's DOM maninipulation methods and Tree Traversal categories with some alterations. One of the main differences is that there is no centralized window.document object to use or reference. Instead, only the original tags provided to tagQuery(tags=) are fully search-able. This difference requires a call to $find(cssSelector) to make any meaningful alterations to the tagQuery() object.

tagQuery() is built to perform complex alterations within a set of tags. For example, it is difficult to find a set of tags and alter the parent tag when working with standard tag objects. With tagQuery(), it is possible to find all <span> tags that match the css selector div .inner span, then ask for the grandparent tag objects, then add a class to these grandparent tag elements. This could be accomplished using code similar to

span")$parent()$parent()$addClass("custom-class")$asTags(selected = FALSE)

This style of alteration is not easily achieved when using typical "pass by value" R objects or standard tag objects.

Tag Query components

Tag environments

"Tag environments" are the building blocks to tagQuery(). When creating a tagQuery(), each tag object is converted to a tag environment. This conversion allows for element alterations to happen in place (pass by reference). Meaning that if a css class is added to each selected tag environment using $addClass() and the result of the method call is ignored, the selected tag environments within the tag query object will still contain the class addition. The added class will exist when the tag query tag environment are converted back to standard tags objects with $asTags().

Tag environments also contain an extra field, .$parent. The .$parent value contains their parent tag environment. The top level tags supplied to tagQuery() will also have a shared parent element. (The shared parent element will have a NULL .$parent value.) This allows for performing sibling alterations at the top level of the tag query object.

The set of tag environments in a pointing to each other within a tag query object can be thought of as a linked list while allowing for a "1 to many" parent to child relationship and up to 1 parent per child.

Tag Query

A tagQuery() behaves simliar to an R6 object in that internal values are altered in place. (but a tagQuery() object is not implemented with R6). The tagQuery()'s methods will return itself as much as possible, unless the method is directly asking for information, e.g. $selected() or $asTags().

Internally, two important pieces of information are maintained: the root elements and the selected elements. The root tag environment will always point (after upgrading to a tag environment) to the original tag object provided to tagQuery(tag=). However, the selected elements are a list of tag environments that update for every $find(cssSelector) call. The selected elements are initialized to a list containing the root tag environment. All tagQuery() methods will act on the selected elements unless declared otherwise.

Tag query objects can be created from other tag query objects. Note, unless there is an intermediate call to $asTags(), the original and new tag query objects will share the same tag environments. The new tag query object will have its selected elements reset. For example:

x <- tagQuery(div())
y <- tagQuery(x)
z <- tagQuery(x$asTags(selected = FALSE))

# Add an example class
y$addClass("example")

# Show `x` and `y` both have the new class
x$asTags()
#> <div class="example"></div>
y$asTags()
#> <div class="example"></div>

# `z` is isolated from the changes in `x` and `y` due to the `$asTags()`
z$asTags()
#> <div></div>

Limitations

tagQuery()s can not be used directly within typical tag locations. An error should be thrown. Instead, please call $asTags() to retrieve the tag structures of the selected tag elements or root element respectively.

CSS Selector

The cssSelector parameter to many methods is a typical CSS selector. Currently tagQuery() understands how to handle any combination of the following CSS selector information:

element:Match against a tag name. Example: div

id: Match against a tag id attribute. Example: #myID

class: Match against a tag's class attribute. Example: .my-class. This may include multiple classes in any order.

The $find(cssSelector) method allows for a CSS selector with multiple selections. Example: div span or .outer > span.inner. All other functions only allow for single-element CSS selections, such as div#myID.my-class.

Methods

All methods return the altered tag query object unless otherwise stated.

Select tags

$find(cssSelector): Find all tag elements matching the multi-element cssSelector starting from each selected element's children. If nothing has been selected, it will start from the root elements. The tag query object's selected elements will be updated with the matching set of tag environments. A new tagQuery() object will be created with the selected items set to the found elements.

$children(cssSelector = NULL): Update the selected elements to contain all direct child elements of the selected elements. If a CSS selector is provided, only the direct children matching the single-element CSS selector will be selected. A new tagQuery() object will be created with the selected items set to the children elements.

$parent(cssSelector = NULL): Update the selected elements to contain the unique set of direct parent of the selected elements. If a CSS selector is provided, only the direct parents matching the single-element CSS selector will be selected. A new tagQuery() object will be created with the selected items set to the parent elements.

$parents(cssSelector = NULL): Update the selected elements to contain the unique set of all ancestors of the selected elements. If a CSS selector is provided, only the ancestors matching the single-element CSS selector will be selected. A new tagQuery() object will be created with the selected items set to the ancestor elements.

$closest(cssSelector = NULL): For each selected element, get the closest ancestor element (including itself) that matches the single-element CSS selector. If cssSelector = NULL, it is equivalent to calling $selected(). A new tagQuery() object will be created with the selected items set to the closest matching elements.

siblings(cssSelector = NULL): Get the siblings of each element in the set of matched elements. If a CSS selector is provided, only the siblings matching the single-element CSS selector will be selected. A new tagQuery() object will be created with the selected items set to the sibling elements.

$filter(fn): Update the selected elements to a subset of the selected elements given fn(el, i) returns TRUE. If fn is a CSS selector, then the selected elements will be filtered if they match the single-element CSS selector. A new tagQuery() object will be created with the selected items set to the filtered selected elements.

$reset(): A new tagQuery() object will be created with the selected items set to the top level tag objects.

Update selected tag info

$addClass(class): Apps class(es) to each of the the selected elements.

$removeClass(class): Removes a set of class values from all of the selected elements.

$hasClass(class): Determine whether the selected elements have a given class. Returns a vector of logical values.

$toggleClass(class): If the a class value is missing, add it. If a class value already exists, remove it.

$addAttrs(...): Add named attributes to all selected children. Similar to tagAppendAttributes().

$removeAttrs(attrs): Removes the provided attributes in each of the selected elements.

$emptyAttrs(): Removes all attributes in each of the selected elements.

$hasAttr(attr): Returns a vector whose values are whether the selected element contains the non-NULL attribute.

Adjust child elements

$append(...): Add all ... objects as children after any existing children to the selected elements. Similar to tagAppendChildren()

$prepend(...): Add all ... objects as children before any existing children to the selected elements. A variation of tagAppendChildren()

$empty(): Remove all children in the selected elements. Use this method before calling $append(...) to replace all selected elements' children.

Adjust Siblings

$after(...): Add all ... objects as siblings after each of the selected elements.

$before(...): Add all ... objects as siblings before each of the selected elements.

$replaceWith(...): Replace all selected elements with .... This also sets the selected elements to an empty set. A new tagQuery() object will be created with an empty set of selected elements.

$remove(...): Remove all selected elements from the tagQuery() object. The selected elements is set to an empty set. A new tagQuery() object will be created with an empty set of selected elements.

Generic methods

$each(fn): Perform function fn on each of the selected elements. fn should accept two arguments: a selected element and the selected element's position within the selected elements. This argument order is different than jQuery's $().each() as there is no concept of a this object inside the function execution. To stay consistent with other methods, the each of the selected tag environments will be given first, followed by the index position. Any alterations to the provided tag environments will persist in calling tag query object.

Tag Query methods

$asTags(selected = TRUE): If selected = TRUE, then all previously found elements (and their descendants) will be converted to tags. If selected = FALSE, the top level tag elements (and their descendants) will be converted to standard tags. If there is more than one element being returned, a tagList() will be used to hold all of the objects.

$root(): Return all top level (root) tags environments. If there are more than one, it will be returned within a tagList(). If there is only one tag, it will be returned.

$selected(): Returns a list of selected tag environments.

$get(position): Returns the selected tag element at the position position.

$rebuild(): Makes sure that all tags have been upgraded to tag environments. Objects wrapped in HTML() will not be inspected or altered. This method is internally called before each method executes and after any alterations where standard tag objects could be introduced into the tag structure.

$print(): Internal print method. Called by print.htmltools.tag.query()